Move to PostgreSQL and improve web UI
Remove SQLite as the server database in favor of PostgreSQL. This allows
refactors in the Diesel models, although the loops approach is still not
removed to minimize changes.

+++ service/src/time.rs
@@ -1,9 +1,9 @@
 use std::time::{SystemTime, UNIX_EPOCH};
 
-pub fn get_timestamp() -> u128 {
+pub fn get_timestamp() -> i64 {
     let since_the_epoch = SystemTime::now()
         .duration_since(UNIX_EPOCH)
         .expect("Time went backwards");
 
-    since_the_epoch.as_millis()
+    since_the_epoch.as_millis() as i64
 }
